feat structure: improved hierarchy of calculators and metrics
