+++ src/cpp_stats/metrics/metric_calculator.py
@@ -1,13 +1,32 @@
-import abc
-import clang.cindex
-from metric import Metric
+from pathlib import Path
 
-class MetricCalculator(abc.ABC):
+import clang.cindex
+
+from cpp_stats.metrics.metric import Metric
+
+class BasicMetricCalculator:
     '''
     Class for calculating specific list of metrics.
     '''
-    
-    @abc.abstractmethod
+
+    def __call__(self, file_paths: list[Path]) -> list[Metric]:
+        '''
+        Calculates metrics for a given set of C/C++ files.
+
+        Parameters:
+        file_paths (list[Path]): paths to C/C++ files
+
+        Returns:
+        list[Metric]: List of metrics calculated for set of files.
+        '''
+
+        pass
+
+class ClangMetricCalculator:
+    '''
+    Class for calculating specific list of metrics.
+    '''
+
     def __call__(self, node: clang.cindex.Cursor) -> list[Metric]:
         '''
         Calculates metrics for a given node that represents function, 
@@ -20,10 +39,9 @@ class MetricCalculator(abc.ABC):
         Returns:
         list[Metric]: List of metrics calculated for node.
         '''
-        
+
         pass
-        
-    @abc.abstractmethod
+
     def observed_cursors(self) -> list[clang.cindex.CursorKind]:
         '''
         Returns list of cursor kinds that can be passed as an argument 
@@ -32,5 +50,5 @@ class MetricCalculator(abc.ABC):
         Returns:
         list[clang.cindex.CursorKind]: List of observed cursor kinds.
         '''
-        
+
         pass
